added get and set servo position commands.  Each servo angle is represented bya  single byte

+++ ros_arduino_firmware/src/libraries/ROSArduinoBridge/servos.ino
@@ -63,7 +63,7 @@ void SweepServo::doSweep()
 
 
 // Set a new target position
-void SweepServo::setTargetPosition(int position)
+void SweepServo::setTargetPosition(uint8_t position)
 {
   this->targetPositionDegrees = position;
 }
